Moved functions and files
