+++ sibylapp/config.py
@@ -18,7 +18,7 @@ MAX_ENTITIES = (
     10  # Maximum number of entities to select from. Set this to None to use all
 )
 DATASET_SIZE = 1000  # Max number of entities to use for dataset-wide visualizations
-LOAD_UPFRONT = True  # If true, run heavy computations on initial load, else greedily run as needed
+LOAD_UPFRONT = False  # If true, run heavy computations on initial load, else greedily run as needed
 
 
 # PRE-WRITTEN FORMAT FUNCTION OPTIONS =============================================================
