+++ sibylapp/view/utils/helpers.py
